LayoutTests/media/unsupported-tracks.html

<video controls></video>
<p>Test that QuickTime file with unsupported track types only generates an error.<p>
<script src=video-test.js></script>
<script>
video.src = "content/unsupported_track.mp4";
waitForEvent("error", function () {
    testExpected("video.error", null, "!=");
    endTest();
});
</script>
